+++ src/fs/reiser4.h
@@ -69,7 +69,7 @@ public:
     }
 
     qint64 maxCapacity() const override;
-    qint64 maxLabelLength() const override;
+    int maxLabelLength() const override;
     SupportTool supportToolName() const override;
     bool supportToolFound() const override;
